Add server::Connection to the public API
