+++ CHANGELOG.md
@@ -5,6 +5,16 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `server::Connection` encapsulates the QUIC connection from a client and
+  provides a protocol-specific connection. This change improves encapsulation
+  and makes the API more idiomatic to review-protocol. Currently it provides the
+  following API:
+  - `send_trusted_domain_list`
+
 ## [0.7.0] - 2024-09-28
 
 ### Added
@@ -234,6 +244,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.7.0...main
 [0.7.0]: https://github.com/petabi/review-protocol/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/petabi/review-protocol/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/petabi/review-protocol/compare/0.4.2...0.5.0
